fixed circular issue
